pyproject.toml
[tool.poetry]
name = "legal-education-platform"
version = "2.0.0"
description = "Educational platform about legitimate consumer rights and legal resources. WARNING: Original version contained dangerous pseudolegal content - all harmful functionality has been removed."
authors = ["Legal Education Contributors"]
readme = "README.md"
packages = [
    { include = "backend" },
    { include = "packages" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
cryptography = "^46.0.2"
fastapi = "^0.111.0"
joblib = "^1.4.2"
langchain = "^0.2.5"
langchain-community = "^0.2.5"
ollama = "^0.2.1"
pillow = "^11.3.0"
pypdf = "^6.0.0"
pypdf2 = "^3.0.1"
pytesseract = "^0.3.13"
pytest = "^8.2.2"
pyyaml = "^6.0.3"
reportlab = "^4.2.0"
requests = "^2.32.3"
scikit-learn = "^1.5.0"
spacy = "^3.7.5"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
werkzeug = "^3.0.3"
pydantic = "^2.0.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
flake8 = "^6.0.0"
black = "^23.0.0"
pre-commit = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[dependency-groups]
dev = [
    "pip-audit (>=2.9.0,<3.0.0)"
]
